+++ CHANGELOG.md
@@ -6,6 +6,7 @@ x.y.z (TBD)
 - The keychain service name used by Realm to manage the encryption keys for sync-related metadata on Apple platforms is now set to the
 bundle identifier. Keys that were previously stored within the Realm-specific keychain service will be transparently migrated to the
 per-application keychain service. ([#1522](https://github.com/realm/realm-dotnet/pull/1522))
+- Added `RealmObject.GetBacklinks` API to dynamically obtain all objects referencing the current one. ([#1533](https://github.com/realm/realm-dotnet/pull/1533))
 
 ### Bug fixes
 - `Realm.GetInstance` will now advance the Realm to the latest version, so you no longer have to call `Refresh` manually after that. ([#1523](https://github.com/realm/realm-dotnet/pull/1523))